Single phase incompressible flow solver
The solver requires access to the hypre library.  This is set through the
variables:
INCLUDE_LOCATIONS
LIBRARY_LOCATIONS
LIBRARIES
in ico_flow/GNUmakefile

This is primarily a proof of concept and so the boundary conditions are hard-
coded to those of a lid-driven cavity.  Data for this test case is generated
via the 'run.sh' script which runs the simulation for nx=ny=32,64,128,256,512.
The data is analyzed and plotted in the 'cavity' notebook.  Good agreement with
the data of Ghia et al is obtained.

Convergence is examined by looking at the min/max velocities along the center-
lines.  With this criteria, the coarsest grid appears to be outside the
asymptotic convergence regime but second order convergence is acheived at the
finer meshes.